edi, eng, renderer: light sens vertex attr (see message)
I don't really like this, but I can't think of any other good solution
for now. I don't want to duplicate materials with modified colors,
because there might be a lot of light sources with same material but
varying intensity. It also to me doesn't make sense to make it part of a
material. Idk, will need to revisit this later.